IMAP  Permit association of arbitrary sender/receiver names
Adds a new addresses.properties file, which contains mappings of
usernames (keys shared with users.properties) to a comma-separated
list of strings which will be matched by the IMAP service when it
checks sender/receiver.

The explicit lists always come from addresses.properties, regardless of
the selected UserDetails implementation in use.  In future, it should be
possible to associate names in other providers (e.g., LDAP).

Signed-off-by: Matt Benjamin <matt@linuxbox.com>