build.xml - Delete entire build directory
The ant build used to output into "bin", and deleting the entire thing
would interfere with Eclipse.  For this reason, only a partial delete
was done when cleaning.  Now, ant outputs into "build/bin", so it
owns the entire "build" directory structure.

Delete the entire "build" directory when cleaning, as clean will
otherwise fail if the "build" directory is not present (for example, a
clean checkout).