Fix locking issues
In the LockService interface, the locking functions were defined as
returning boolean indicating whether or not the lock was aquired, and as
throwing an exception sometimes when the lock failed.  Not every caller
was checking the return value; some were depending on the exception,
which is not sufficient.

Modify this interface, it's implementors, and callers to return void,
and always throw an exception if the lock fails.  This way, a caller
cannot access protected data unless they have the lock.

Modify the locking unit tests to handle the New World Order.

Add an integration test to stress locking.