MailPollClient - Add the ability to poll for mail from a mail server
This adds a new class of objects to Enkive - EnkiveClients.  These are
similar to EnkiveServers, but are intended to be single threaded and
proactive, rather than multithreaded and reactive.

Within this class is an AbstractPollClient that implements the core of a
poller.

Extending this is a MailPollClient, which can poll for email from a
configured server.