shutdown - Clean up CallableProcessExecutor threadpool
TaskPoolAsyncMessageSearchService creates a CallableProcessExecutor
(which is a ThreadPoolExecutor) to handle searches.  However,
TaskPoolAsyncMessageSearchService had no shutdown() method, so it was not
calling the shutdown() method for CallableProcessExecutor.
This left threads running forever if a search had ever been run, causing
shutdown to hang.

Add a shutdown() method to TaskPoolAsyncMessageSearchService that does
the right thing.