Handle RFC 5987 encoded UTF-8 filenames
RFC 5987 defines a way of using non-ascii filenames in a
Content-Disposition header.  This involves using

filename*=<encoding>''<filename>

However, mime4j does not handle these, or even recognize them as a
filename parameter at all.  As a workaround until a better solution is
found, look for a "filename*" parameter if no filename exists, and
decode and store it for later display.