Disable search folders
Search folders are an incompletely designed and unused feature, and are
extraneous at this time.  Disable use of them in the code until such
time as they can be properly integrated.

Since search folders result in a copy of every saved search and an
additional Collection in Mongo, this will reduce DB use and size