Change structure so SearchQuery contains SearchResult
Until now, Enkive searching operated on SearchResults.  Results were
static, and never updated.  Each search created a new results document
in the DB, and those results were frozen in time.  Searching returned a
result, and all operations were on results.

The plan is to make actual results ephemeral, stored in a capped
collection as a cache.  To do this, everything needs to be changed to
operate on queries, and use those queries to get their results.  This
actually simplifies the code a bit, since it gets rid of almost all
ResultBuilders.

Document the new search heirarchy

Add migration for query <-> result

	Bump the DB version to 2 (will change more later), and add a
	migrator for migrating the DB to the new query/result framework.

	- Remove search-folder fields from workspace
	- Replace SearchResults list of SearchResult IDs with
	  SearchQueries list of SearchQuery IDs
	- Move metadata from SearchResult to SearchQuery
	- Drop the SearchFolders and SearchFolderSearchResults collections