Add a monotonic ID to messages.
Message UUIDs are a has of the message, so that they can be
de-duplicated in the database.  This means that there is no way to
determine arrival order, which is useful for updating queries and for
IMAP mappings.

Add a new indexed ID to messages that monotonically increases as the
messages arrive.

- Have search results return a Map of <monotonic_id, _id>
- Store the last monotonic ID in the query
- Have the query use that when re-searching
- Add Index on Monotonic ID + _id
	- Order is important, as we will want to search on just
	  monotonic or on both, and _id is automatically indexed.
- Add DB Mitgration code to add the monotonic ID to messages on 1 to 2
  migration