Bug 15 - Don't buffer exports in memory
Because of the way that Surf architects it's web scripts, it's not
possible to get direct access to the stream of data coming from the back
end.  This means that, if we try to copy the exported mbox to the client
browser, Surf buffers the entire thing in memory first.  This is fine
for small searches, but quickly exhausts the heap available to the web
script for large searches.

Instead, have Enkive export the search to a temporary mbox file, and
send that file to the client.  This means that in-memory buffering is
minimal for the whole chain.