Bug #19 - Preserve sort order when paging
Before, only one URI, the base one, was stored in the model, so paging
always updated to the base sort order.

Add a second URI that includes the sort order, and make paging use that,
while sorting still uses the base URI.  This means that paging continues
through the current sort order, but changing the sort order resets to
page one of that new order.