Search results - show useful error to user
Add a framework for allowing error messages to propegate all the way
bace to ediscovery, and be displayed to the user, even when AJAX is
loading a sub-section of the page (as when jumping to the next page of
search results).

Add a useful error when Mongo can't sort a query due to it being too
long.