FileDocStore - A document store based on files
As attachment sizes and numbers grow, GridFS write speed has become a
huge bottleneck, causing speeds too slow to use.  Create a hybrid
mongo/filesystem store called FileDocStore that stores the Documents
themselves as files in the filesystem, and the metadata for them in
MongoDB.

This involves:
- A new DocStoreService
- A new convenience class
- A new unit test
- A DB version bump
- A migration from MongoGridFSDocStore to FileDocStore

The result is an ~4x speedup in import speeds for Document-heavy emails
on stock ext4.

This also bumps the main Enkive version to 1.4-rc2