VersionCheck - Don't fail if URL returns no data
There was a bug where the version check would fail, aborting startup, if
the URL did not return any JSON (as opposed to being down entirely).  If
this is the case, gracefully fail the version check.

There was another bug where strings (including the version string) were
not properly URL encoded.

Finally, swich the version check to https.

Escape strings